Unified _list/substitute and _list/tabulate.
As long as we don't need to tabulate over multiple keys per document, we might
as well use one list that handles linked docs and key/value pairs. The best
part is that views don't look esoteric anymore. The just read
_list/tabulate/<what>?<query args>. Also I excluded query arguments of _view/
to be excluded from parsing. This fixes most of the errors from before.