Improved reduce function of _view/types-fields.
If group_level is set to one, the new reduce function yields a list of every
field that was ever used in this specific type of document, as given by key[0].
This should make it easy to achieve some kind of 'document-driven design':
instead of specifying every freaking form field beforehand, we just look into
what's already being used and create a form from that. And perhaps leave one or
two open fields if the user wants to define something new.