Imposed restrictions links to other documents.
1) type must be indicated such as '{.., Gemarkung:{_id:"abc"}, ..}, this means
  only known types are allowed,
2) _id is the only allowed property and
3) it must be a string
All of these are useful but I also might be able to live without all these
restrictions. The reason for 3) is that all doc._id's must be strings, so links
to other docs should as well. 2) is meant to avert the user from stuffing extra
information into links that do not belong there. 1) would allow me to display
the right link/graphic button without an extra httpget.