Found a nicer way to work around erica's push restriction.
As introduced in commit facca2f19, we might as well place a ; at the beginning
of a line, so it doesn't match the ill-designed regexp of erica, that strips
names from every function regardless of scope.
As soon as I got the password to my github account and filed the pull request,
my javascript will look nice and clean again.