Adapted a some files to good javascript practice.
Auto-declaration of variables is one serious design flaw in javascript. It
leads to programmers not using the 'var' keyword and then accidentally
overwriting globals at the end of the day. The best practice, of course, is
using 'var' whenever a new variable needs to be defined. Even if we could do
without it. This applies to for-loops even more.