Installed substitute into updates.
The essence of NoSQL is not to use reference tables at all. Whenever a
reference table is introduced, the raw data ceases to be self-explanatory and
gets unreadable. One "good thing" about reference tables however, is that we
have to change the referenced value in one place and it then applies to all
documents. The same thing can be achieved by a sed-style edit of the entire
JSON structure as given by /_updates/substitute. The syntax reads
  /_updates/substitute/<doc>?/<pattern>/<flags>=<new string>&<...>