Added a main page from which one can log in and out.
Wow look at me go!!! The plan would be to stick pretty much to how it's done in
futon, because I like its efficiency best. And as long as we don't go overboard
with the dynamic content, layout and so on, we should be able to run this on a
variety of devices.