Added a neat replication inferface.
Unfortunately it is not that easy to setup CORS in a local couch
installation. So in case we are at the central app, I provided links to
- couchdb.apache.org to install CouchDB
- local replicator futon page to fetch the central app
- local replicator interface.
Also it is impossible to push to the local couch, so replication has to be done
from localhost anyway. When in the local replicator interface, there are buttons
for push and pull requests and we also see response messages.