Scratched the geojson reduce function.
Due to the size limit of reduce's return value (900 bytes or something) I am
unable to use it to collect docs in it that share a GeoJSON object. I might
turn it off by posting to /_config but then I'd have to do so in every local
installation (and assume admin party etc. NO WAY!!!!).