Designed a draft for a spacial _list/bbox-filter.
This one accepts parameters as request body, because I won't use it much with
curl when I don't need to test it. I.e. it is POST _list/bbox-filter/geojson
that also sounds like a sentence. I see that the filter might be switched, but
the geojson part is pretty much set in stone. Parameters are:
- bbox: four-element array [x1,y1,x2,y2] of min/max coordinates,
- limit: stop filtering after this many GeoJSON items,
- keys: array of GeoJSON ids to restrict filter to. This is of course just
  CouchDB API, but I foresee it as as powerful tool to lower the server load by
  providing previously sorted GeoJSON ids.
Output reads {"doc_18b":{"docs":[{"type",etc.}, more docs],"GeoJSON":{...}},
more ids} so it is the GeoJSON id, then docs and GeoJSON, then the next id.