Added type 'Flurstück' as replacement for 'FlurStk'.
For now both are valid types but I will remove the latter in the near Future. I
am sick and tired of people working around special characters they would
normally type in their language, if all we have to use is UTF-8 and that's it.
I won't even refrain from putting ÄÖÜ etc. in my js and commit messages!!!!!