Rewrite statement block as lambda in _list/tabulate.
I noticed this when I was working on the Douglas-Peucker implementation: the
ability to jump to the end of the block with some meaningful result enables me
to express this more concisely. Even with the arguably obfuscated closure, I
rid a bunch of verbose statements that have been there solely to say "ignore
this next block but still do the doc={}".