Refined Douglas-Peucker error metric.
The old metric pretended all lines where infinitely long where in fact there is
a difference between a point being close to the line segment of just close to
the line. By 'line' I mean the line segment continued beyond the points
defining it. There is a nice ASCII sketch describing the idea.