Introduced delays between POSTS and in update_geometry().
We don't have threads in javascript, so any huge geometry update or ajax
callback causes the map canvas to freeze for a couple of milliseconds. I have
imposed the following restrictions on the map interface:
- wait 500ms between geometry list updates
- wait 50ms between two single geometry refreshs.
I've pulled these numbers out of my nose. The new experience on my LenovoX60s
is as decent as it gets, on FF24.0. It kicks ass on Chrome (as expected). The
laggy update looks kind-a retro, which is cool, and may be switched off in a
form as soon as I get to it.