Implemented text field fo polygon count and buttons.
The buttons are labeled 'decrease' and 'increase' but they certainly DO NOT
JUST INCREASE THE POLYGON COUNT BY ONE!!! BECAUSE NOBODY WANTS TO CLICK TEN
TIMES TO GET FROM 12 TO 22 JESUS CHRIST. Increase sets a limit of e.g. 22 to
50, then 100, 200 etc.
The implementation of both buttons looks fairly unelegant but I don't see how
to do this with less lines of code. They look similar but nevertheless differ
enough to remain separated.
Labels are also misleading because the field should say 'object count' or
something.