Whipped up a revision where CouchDB crashes constantly.
This is completely beyond my understanding, but it seems that
_list/bbox-filter/bbox-size manages to succeed but _list/bbox-filter/geojson
then produces a kind-of internal jQuery error that does not describe itself
very well. We end up with no shapes set and the application stops running, so
this makes sense again. So a nice way would be to
- expect shapes to be missing
- continue with calls if $.ajax went wrong.
This works in FF but Chrome still halts but at least without any JS errors.