Suppressed unneccessary keys from filter query.
The client app now uses a stack of cascading viewports, each one contained by
the one before. These also yield an index of where to resume the 'keys' array,
and a 'prefix' (again), that states which items are inside this viewport.
The results are as follows:
i)  I am able to skip over a vast number of bboxes that where
    outside my viewport anyway.
ii) The slowly degenerating prefix/sorted composition approch from commit
    08a2fa635 is no longer used. This is good news since I have no idea how to
    remove items from prefix back to sorted anyway.