Removed 'unsorted' from, and simplified _list/bbox-filter.
If keys are always sorted anyway, there is no need the 'unsorted' query
parameter, and we don't need to binary-search-and-insert-sort the retrieved
items list (although that looked pretty cool). It might even absolve the
CouchDB crashes.
It does not behave well if limit==0. Still this is way nicer than before.