Limited posted keys if viewport has fewer items than requested.
This is nice because bbox-filter now doesn't have to go through the entire list
all the time. The reasoning however is simple:
As soon as there are less items reported to be in the viewport than the app has
actually asked for, it is safe to say that the filter has gone all-the-way
through the list - and this must be it. So as soon as we are in a nested
viewport, we may only specify the keys in prefix.