Moved query params into an 'options' object.
I do not like this at all because as soon as I want to debug _list/bbox-filter
with curl, I have to deal with horrible URI encoding. Since CouchDB complains
even if I used 'count' instead of 'limit' (because it used to be count or
something), I have to nest all parameters. I don't know how else to name it!!!