Revised construction of 'paths' object required by Maps.Polygon etc.
GeoJSON's coordinates have the same structure as the paths property passed to
Google Maps' polygon/polyline/marker constructor. It makes more sense to just
replace the innermost [lng,lat] with a LatLng(lat,long) object and leave the
rest of the structure as it is.
The for-loops look ugly on purpose because they have to be fast.