Added lib/range.
Because I need a small utility to test if a given date falls inside a time
range or if two time ranges overlap and so on. It should handle dates as
strings such as "05.5.2012" or just "1992" or "5.1999-2001" etc, so if the day
or month is omitted it should assume that the whole day/month was meant. Years
still have to be four digits and it does not care about leap days or how many
days a month has.
I found this waaay harder to write than I thought. Testing was a nightmare!!!