Added _view/geojson-doc.
This probably gets more and more confusing because there is already
_view/geojson and _view/bbox. The latter is also capable of mapping documents
to GeoJSON keys (this ability won't be needed in the near future).
It's all starting to make sense as soon as geometric and content-based are
clearly separated and we have a handful of views each designated to only one of
those two tasks.