Added unified time range to _view/dimensions reduce output.
CommonJS is unavailable for reduce functions so I had to quote it from
lib/range. Avoided the use of range/toRange or range/toString because they are
most likely to be changed for more flexibility and I would end up with more
code maintenance. The expand function could look simpler for sure.