Used 'setOptions' to change polygon geometries.
I deem it impossible that polygon reduction would lead to a different polygon
type. That means I can just use 'setOptions' to exchange geometry (instead of
deleting and recreating a Maps Polygon).
This is presumably faster (it should be) and I don't have to store fillColor
anywhere. This property would be lost otherwise because I don't want to query
the document type twice.