Implemented content-filter listing.
This looks great and the best news is that it *almost* works. Content-filters
are stacked in a handy list and may be removed at any time (this is done by
setting cut.remove_me=true and pass this to filter_documents(), this way I
could re-use a vast amount of its logic).
I haven't found out yet why filters are occasionally left in the queue but it
is CLOSING TIME anyways.