Added _view/fields-values.
I will expand this to do some recursion into objects but right now this is
completely sufficient to assist with some auto-completion.
Providing an end_key is kind-of a bitch; I figured appending a space to the
last start_key element and using inclusive_end=false should work most reliably.