Enabled field-selection when a filter is added.
I made everything most useful as such:
1) if field and value is given, the property must match
2) if no value is given, the property must exist
3) if no field is given, any value must match.
I still have stupid labels inside the script part and this is going to be a
nightmare if noone fixes it. Also the .change() hook does not really behave
like I want, because now a filter is added as soon as I changed the text input
and left the box. This has the potential to be very, VERY annoying.