Improved type-selection form.
I am a horrible person. No one wants to de-select type after type until only
the one or two of interest remain.
As a bonus, I listed the type restriction as with any other filter in order to
reduce confusion over if the filter is applied or not right now.