Changed type filter to use the #type-is <li>.
This also includes a small but sensible change in how the form behaves: if no
type is selected, the filter is not disabled. Otherwise it would result in an
empty set like it had been before. The #type-is list item vanishes depending on
the filter being in effect or not.