This evades duplicate columns.
One small glich remains, that, if I delete a column there is no check if that
one is even available by the current type selection. I just moved it to the
other <optgroup> without any verification.
I definitely can live with that, given the useless additional amount of coding.
The code looks complicated enough as it is already!!! WHO ON EARTH WOULD TRY TO
UNDERSTAND ALL THAT????