Improved _view/fields-values.
This new view makes more use of the reduce function by providing a count, a
range for numbers and if the field happens to be an object or array, I can
probably think of something too.
'doc.type' was moved to the end of the array key because I want to histogram
identical field names of different doc types.