Simplified _list/tabulate API and map.html.
Reverting the state to the form just before the new filter was issued went more
and more complicated and also the use of an <option value="_expression"> was
confusing. I moved cut.expression to cut.value where it made more sense, and
introduced special cut.field identifiers '_expression' and '_keyword'. Those
coincide with my option values on the map page.
I'd still say that '_keyword' is a bit misleading; it is just an ordinary
RegExp that is applied to the full document. Same goes for '_expression': it
does not imply "JavaScript expression" and may easily be confused with "Regular
Expression" or something.