Preserved order of fields as given by _view/types-fields.
The technique behind this is that for-in-loops yield property names in the same
order as new properties are added to the object. Even if properties are all
'undefined'. If they are re-assigned later, the order is still holds.
Also, JSON.stringify() skips over 'undefined' properties (as it should) and
thusly a presumably complex task gets done with relatively little effort.