Initiated autocomplete-initalization on page load.
As the comment says, FireFox 26.0 does not clear or reset any form elements if
the page is reloaded. So the type may already be set to a valid one but no
field info is displayed. The user would have to assign a different type and
then the first one again. HOW GHASTLY!!!!