Provided standard doc[doc.type] description field.
For instance I expect documents in map.html to have this kind of document
description and I cannot for the life of me think of any case where such a
field would make no sense. I therefore encourage the user to provide one as
well.
Some shape files do not come with a description (it may just be part of the
file name etc.), so the new form applies the same for all documents.