Added 'select similar geometry' map form.
It will suggest similar geometries to the user because in my experience there
is completely no need whatsoever to add a second rectangle that is just
slightly offset to a first one. On the contrary: map readability and
comprehensiveness are greatly improved if data is attached to the same shape
when it obviously belongs together.
Re-using geometry is not supported by shapefiles (or GeoJSON), so we have to
check the database and the GeoJSON itself.
I cut short the latter step by uploading the data temporarily: polys need to be
assigned using the map form, then revert the upload and upload again. This
leaves some room for improvement.