Exchanged 'doc' for 'feature' as member of the features array.
Also I checked for the presence of 'shape' in a more detailed way, so that js
doesn't throw exceptions anymore.
Unfortunately I don't see any way to provide the (WGS84-transformed) new
geometry after the documents have been uploaded WITH the GeoJSON_clone-member.
This means the user cannot revert-back the decision and the geometry stays
cloned. I could cache the shapes but meh....(There is probably no need to fix
this as long as noone complains. I see this as a minor problem.)