Added _list/export to XML/GeoJSON/whatever.
In contrast to my previous beliefs, ArcGIS manages to read .geojson-files just
fine as long as I have a top-level 'name' property set. That pretty much
settles my needs for geometry export.
MS Excel as well as LibreOffice cope with a stupid bloated XML scheme. At least
it's not binary, so I'm going to cover that as well.
Output will be filtered by listing _view/geojson keys as encoded URI. I am well
aware that the amount of these is limited to ~50 or something, so as soon as
they are exceeded, I have to come up with a compressed scheme that is handled
inside the list.