Added compressed_keys query parameter to _list/export.
My conundrum that I have with forms and CouchDB renders as such:
- Pure HTML5 form posts are very much unable to create JSON-compliant payload
  data. Neither with hidden inputs nor with $().submit() hooks.
- I want to point the browser to something it can save as a file.
- Data URI Scheme/Blobs have limits in their implementation which makes them
  not that safe-to-use.
That leaves me with GET requests solely.
Therefore I have to cope with limitations on the amounts of keys that can be
specified in the request params, 8.1k or something.
This justifies the hacky compressed key scheme I came up with: it enables the
list function to manually print-skip-print-skip-... the given number of docs.
I used a top-level query parameter in favor of a nested 'options' object,
because I still want the URL to be kind-of transparent.
I used the dash because it doesn't get translated into a %2Fetc-encoded string;
this saves space, as I am unsure how far I can go with this.