Added xml table to _list/export's file types.
I employed some very crafty single-instanciated-lookup-objects to initialize
the response header without making too much fuzz. If no fields are given, the
list yields all there is, with optional geometry id and revision as the first
columns. A column will be rendered for each specified field, even if we ended
up with an empty column.