Fixed similar geometry search in combination w/GeoJSON_clone.
I should probably delete the entire file now because it looks like a convoluted
mess. Who on earth is supposed to understand this??
The (pretty useless) functionality makes sense on the other hand because even
if geometry was cloned, the map is suggesting geometry that looks like the
cloned one.