Fixed update of existing document fields.
Half of these are plain stupid coding errors and the other half did not respect
existing fields of the previous document. The priority should always be
new-value > existing-value > undefined.
I realize that it is kind-a difficult to un-set doc.type and doc.time from
within the form (this is probably a good thing).
GeoJSON feature geometries are ignored if a GeoJSON_clone is found either in
the uploaded or previous data. As with the other fields, it can be unset with
an explicit empty column.
The 'revert' button just deletes all the documents right now and is another
story.