Detached doc from uploaded_docs[id].
This is not the best way to do this, but a simple grep will show me all
occurrences of this inferior pattern, to be fixed in an extra commit.
I need those two detached because doc will be modified according to the
uploaded data and I want uploaded_docs to store the previous version of it.