Changed invalid uploaded_docs[d].id with feature.id.
The good thing is that feature.id is not part of feature.properties, so it goes
undetected if documents are uploaded. The other good thing is that I don't need
the stupid index anymore and I end up with a simpler script.
The revert button is still not working correctly, however!!! I am in the middle
of if!!!!