This is a work-around for duplicate shapes in upload.html.
It is the best I can come up with at this late our. It fixes duplicate shapes
on the geometry selection map canvas that are left uncleaned if the user skips
to the next feauture. As far as I can guess, they are produced by the following
combination:
User updates docs and has GeoJSON_clone set explicitly to the document id (to
inform the application that no update of the Geometry is needed). He then
decides to check for similar geometry.
In this case I accidently queried for [GeoJSON_clone,0] and identically for
[feature.id,0] and stored my geometry twice in feature.items[GeoJSON_clone].