Changed exported CRS to EPSG:3397.
Quantum GIS' on-the-fly CRS transformation seems to work in the worst possible
way against WGS84: Sometimes the polys are placed correctly in WGS84, sometimes
not, and when in any other CRS they are always messed up. Gauss-Kruger zone 4
(or maybe zone 3) causes seemingly less trouble.
Of course this will get read as a query option in due time.