Fixed a very very small glitch in the map page.
If a table column was changed, I staged every item to update its documents, but
the geometry was not explicitely un-staged for an update. Now if #delayed was
checked, the table seemed to be updated in an unneccessarily delayed fashion.
And it created some redundant traffic by fetching unchanged geometry.