Removed unsignificant digits from errors listing.
This way the geometry lookup using _view/geojson is more reliable because I can
thrust the error given by _list/bbox to be correct.
The raw view output still reads "errors": [0.0004299999, 0.00017000001, etc.],
because it seems that numbers are converted to floats, then to a string
representation, and then added to the index.