Improved autocomplete to use text input as regular expression.
The previous approach was kind-of half-baked because it completed to regular
expressions (i.e. escaped labels) but did not match against them.
Now anything typed in the text input will be treated as reqular expression and
I can see instantly if I made a mistake and what the filter results are.