Added _deleted as special field indicator.
This is because I have to check the link integrity before posting the documents
in bulk, i.e. when either #revert is pressed or in load_features(). While
unchecking the _deleted property would leave the database intact, I'd end up
with duplicate geometries.
Admittedly this is a very poor way of solving it.