Sensibilized pointInPolygon() for points close to line segments.
The commit looks like a mess because everything needed review.

For the signature with known_point, vector cross products are written to
operate with p-q, p-a, a-b, b-q in that order:
- p-q is constant,
- p-a is used to check if a is close to b,
- a-b is then used to to check if p is close to a-b. The actual distance
  (a-b) x (p-a) / |a-b| was approximated by using abs(ba_u)+abs(ba_v) instead
  of a square root. This is faster and the error is within 42%.
- Last b-q is needed to check CCW(p,a,b)!=CCW(q,a,b) && CCW(a,p,q)!=CCW(b,p,q)
  expressed as (p-a)x(a-b) == (a-b)x(b-q) and (p-a)x(p-q) == (p-q)x(b-q).
For the signature without known point, I removed a bunch of precautions for
unlikely cases and worked around comparisons with absolute zero.