Reduced amount of pre-filter keys for the 2nd reduction.
At the 2nd reduction neither time nor types are carried into options
necessarily, so the pre-filter has to look the missing one up in the filter
chain.
Now, because I've allowed to branch into pre-filtering only for the 1st and 2nd
filter, the actual look-up-process just has to go through filters[0] and [1]
and this is why it looks as simple as that. My assumption has always been that the
amount of documents must have reduced sufficiently after that.