Fixed uncaught exception thrown in filter_documents.
This happens when spatial filters clean up the filter queue by using {union:[]}
as the first argument. This causes all filters to be collapsed and the
collapsed filter runs by _list/tabulate. If it contains either time or type,
pre-filtering is used (as it still should), but it stacks {redo:filter} on top
of the argument list. filter_documents then gets confused because it processes
{spatial:relation} before {redo:filter}.