Improved reduce function of _view/type-year.
It works now in conjunction with _list/tabulate because row.value has a valid
document in it. This is neat because I can move the pre-filter key selection
away from the map page, as it is just now.
The value is set by the reduce function exclusively, because the map page still
has to run through all these rows with reduce=false, and it isn't needed there
at all.
I wish I could reduce to a compiled range instead but this would
require('views/lib/ranges') and those aren't allowed in a reduce function, so
this is as good as it gets.