Changed update_geometry() to update geometry and then docs.
I have trouble finding out what problem FF31 has with my page because the user
experience is still very unpleasant. My approach is to group updates to the
table and updates to the map canvas to reduce the amount of re-flows.
However, even if I remove the table, the page is slow and laggy.