Absorbed coordinate offset into utils/proj4().
I can't decide if either QGIS or ArcGIS transforms to WGS84Vcorrectly, so I
just refrain from importing any data in this crs, as I always did before. The
offset is now neatly cached into the views and doesn't have to be added at
runtime anymore.
The function has the same effect as to require('./proj4') directly, but if the
signature proj4(definition) is used, it wraps forward/inverse and applies the
coordinate offset from script/google.maps.js.