Reviewed proj4js projection definitions AGAIN.
Ok, now it seems that at least QGIS and proj4js have the same understanding of
the towgs84 parameter set. It just should have been added to all of my
projection definitions, because QGIS seems to apply it all the time.
Coordinates remain displaced to Google Maps aerials, but not as far as before,
meaning that lib/proj4() has to stay.
On another positive side note, the shift in exported coordinates reduced to
0.7mm. I attribute this to coincidence or something.