Redefined 'intersects' to distinguish polys and linestrings.
What bothered me with the old definition was, that polys have been
'intersecting' with other polys when they only shared a boundary or a single
point. It is mathematically counterintuitive, because the intersecting part in
this case should be an actual two-dimensional object.
When considering polys and linestrings, however, the intersecting part has
always one or zero dimensions, so no distinction should be made whether it
coincides with the boundary or not.