Simplified signature used to continue tail recursion.
I just could as well use recall_arguments instead of recall_arguments[0], if I
need a tag to tell filter_documents to go on with the tail recursion.
However, filter_documents isn't called without any arguments at the very end,
so the part where the viewport is updated had to move down, where new_filter is
set (or not).