Changed upload page to interpret 'GeoJSON_cl' as GeoJSON_clone.
Field names in ESRI shape files are limited to 10 characters. Much to my
dismay, I am forced to work with this depraved and useless format from time to
time (for instance when 'joining attributes by location').
Then of course the last three characters of GeoJSON_clone are dropped. I could
rectify this in the upload page myself by entering the correct field name but
this is faster and anticipates the likely use-case.
ESRI I hate you so much. You are the worst of humanity.