Fixed wrong order of autocomplete suggestions.
I've just made sure that the key is a string by prefixing it with an underscore
and restored the original string before matching is done.
I am relying on the preservation of insertion order of object keys here,
despite that the ECMAScript standard explicitely states the opposite. However,
I feel I am allowed to do this because in this case the order is absolutely not
critical.