Changed span.types to be filled with <label>s.
The space is needed to allow line breaks after each type checkbox. That makes
the entire line a little bit broader. On the other hand the grouping makes more
sense this way.
I used $.text to initialize the label because otherwhise everyone with
write-access wants to fill it with script tags. Not on my watch.