Crafted a rudimentary animation for 'filter in progress'.
It waits 0.5 seconds until it is displayed, because most of the filters are
effective immediately, and in that case it is more distracting than telling the
user to wait patiently.
On a lighter note, even with spatial filters on vast geometry selections I
haven't waited more than perhaps six to eight seconds for the filter to
complete. On my Lenovo X60 that is. So this is just in case.