Prefixed target with 'http://' if replication was unsuccessful.
I can't assume a wrong target if it contains a slash and it has no http or
https at the beginning, because it still may constitute a valid local target on
CouchBase or some other implementation.
So I've let one replication deliberately run with the specified target, and
only if this one failed and the above conditions are met, prefix the target and
post to /_replicate again.