Formatted a nice tool tip with document contents.
This is more of an exercise to do it with just one assigment. I will remove
that horrend bit as soon as the tool tip is annoying me.
Also I probably shouldn't store an entire document in each title attribute.
This should be loaded dynamically and only if it's needed.