Fixed relay_geometry to work with multiple GeoJSON clones.
The old version most likely just replaced multiple geometry keys with a single
new one (I've never tested this), thus damaging the entire document set.
So I've rewritten the function to load missing documents first and then replace
the id inside GeoJSON_clone with an appropriate new one:
- If we have a document that links to the relayed geometry exclusively (i.e. it
  has only one id inside GeoJSON_clone), we can safely move the geometry into
  it.
- Otherwise remove type and _deleted from the relayed document. This makes it
  invisible in the map page, but the geometry is still accessible to documents
  linking to it.
The above two are only needed if the relayed document is about about to be
deleted from the database.
In any case I switched to replacing the old id substring with the new one.