This simplifies control flow a bit.
I wasn't particularly keen on the return; out of the nested if() statements;
Neither on 'single' being pushed into relays as if it shouldn't have been
popped out of relays in the first place.
Also the redundant boolish !!single in the 2nd if() statement is written out of
sheer need for clarity. No one wants to read this horrible mess, AND ON TOP OF
THAT likes to deliberate on how the 'return;' already rules out !!single.