Renamed 'single' to 'acceptor'.
Because the feature will function such as it will accept the geometry from the
deleted feature as its own, and we can't move the geometry if we don't have an
acceptor for it, etc.
This isn't quite the term I was looking for but I come up with anything better
than this.