This is how the loop was intended from the very beginning.
I may only set r to its predecessor if the preceding key is identical to this
one. r-- evaluates false-ish only if r===0, but in this case the preceding key
is undefined anyway, so the lazy evaluation won't reach the statement.
The other problem is, of course, that it is a waste of time for anyone that has
to read though it.