Draft of a query to _spatal/bbox if GeoCouch is installed.
_spatial/_list doesn't allow POST, so there is no way to sieve matching
geometries by a list of document ids, (as it is done in
_list/bbox-filter/bbox), because I can only send something like 200 keys to
_spatial/_list in a GET request.
The bottom line is that I'll end up with a query for ALL ids (which is
disgusting). To make it less painful, I excluded the geometry errors from
_spatial/bbox because I'd have to transmit and JSON.parse all of them seems,
even if only 50 are needed.
Geometry keys that intersect with my viewport are sent to a 2nd query to
_list/bbox-filter/bbox, but in this case with only options.error set. That
allows me to reuse the success function.