Pre-filtered geometries with _spatial/bbox for spatial relations.
This is similar to how it's implemented in update_items(): All matching ids are
hashed for faster look-up, and in a second step, options.keys are run by that
hash-table.
_list/bbox-filter still needs to be queried twice, because _spatial/bbox
returns intersecting bounding boxes. It shouldn't hurt to further reduce these
to contains/within before running the extensive point-in-polygon query.