Sorted input files into ogr2ogr_uploads object.
Hence ogr2ogr_uploads is structured as { <basename> : { <extension> :
<content>, etc. }, etc.} and content is a base64-encoded binary.
I assume that everything that was sorted into the same basename belongs to a
shape file blob. This makes it impossible to upload for instance my-data.kml
AND my-data.sqlite at the same time, and then expect the upload page to
recognize these as two different feature collections.
It's not my fault, it is ESRI's fault. They came up with a stupid concept of
having a file for (not even spatial) index and for the dbase and for geometry
and for the projection definition.
ESRI you've ruined the world for everyone!!!!!