Sketching criterion for pre-filter to _view/fields-values.
I've included _keyword as well as field/value because even _keyword might run
faster in some cases if I crafted this in an awesome way:
- Document counts of less than 100 won't be pre-filtered.
- _expressions are no use to pre-filter in any case.
- Single fields over multiple types might not be good to pre-filter; I suspect
  _view/fields-values to run over too many unused rows. On the other hand if
  you have a bunch of very similar documents, who knows?
- If the mere existence of a field has to be known, group_level expands over
  [<type>, <field>] to improve _list response time.
Regardless of these precautions, it is still essential ensuring the number of
rows processed by _list/tabulate to not exceed the number of documents.
However, because I don't see a safe and easy way to count the number of unique
keys, I placed a limit of 1 and skipped over the number of rows I set as a
threshold.