Changed update_items to extend the last viewport.
The last viewport can only be extended if only one vp was removed from the
stack and if the list("bbox-filter/bbox") returned the same items.
I still could construct a scenario where these conditions would result in a bad
viewport, perhaps if limit=1 and we have a large item and a larger item in the
middle of the first one. The larget item would have to cover lots of vertical
space but almost no horizontal one. Then the user has to move the map really
fast from one side to the other. Etc.