Added pad to the viewport stack.
pad is a bit tricky: it touches the most outer border of all the displayed
geometries' bboxes borders directing to the center. So, as soon as pad lies
outside of the viewport, we can assume the same for one or more of the
displayed geometries.
Inversely, we can skip the geometry update as long as it's inside.