Preceeded _spatial/bbox query with option count:true.
I am not sure yet if this helps. It obviously lakes workload off the client but
asking for the amount of geometry takes about as long as asking for the
geometry itself.
There should be a viewport array for this to avoid querying crowded areas
twice. Unfortunately this one would have to be build from inside out, so it is
inversely to the one already in place.