Added multi-key-fetch to autocomplete suggestions.
I noticed that the average user takes a gazillion years to type in a regexp.
While he does that, the application has plenty of time to sequentially load all
the values, and no latency whatsoever will be noticed.
Inversely, when the pre-filter reverts to _list/tabulate/field-type-value, the
perceived time between pressing <enter> and getting a result seems infinite.
So what I can offer is that the user may as well use the results already
filtered client-sidely, as an extra autocomplete suggestion. It will be
displayed as soon as less than 50 items are suggested, but it hides if there's
only one other item.