Included union in _view/type-year.
If queried with group=true, the value changes to an object with the property
'one'. 'one' is a range-object that covers the time ranges of each document.
So, if it happens to translate to "07/2009-06/2010" and I want to have
documents from 2009-2010, I don't have to check each document individually.
If I ask for documents of 06/2010, however, the documents of this key must be
checked because there might be one from 09/2009 etc.