Switched 'tilesloaded' and _view/geojson-doc callbacks.
The rationale would be that I don't have to wait for the map to have loaded
completely before I may issue expensive calls to _view/geojson-doc and
_view/dimensions.
As the API documentation states, !!map.getBounds() indicates if the map is
initialized or not. So I used it to decide to stick finalizization on
'tilesloaded' or to run it directly.